[feat] AWS S3 추가, S3Service, [fix] 사용자 마이페이지 차량 정보 리스트 추가

+++ carparter/src/main/java/com/spring/carparter/entity/UserCar.java
@@ -24,7 +24,7 @@ public class UserCar {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "user_car_id")
-    private Integer userCarId;
+    private Long userCarId;
 
     /** 차량 소유자 (User) */
     @ManyToOne(fetch = FetchType.LAZY)
